file relocation, col name changes
